#359: Simple lift and shift
