+++ lib/Builder.js
@@ -3,6 +3,7 @@
 const TYPE = require('../types/mcdev.d');
 const Util = require('./util/util');
 const File = require('./util/file');
+const config = require('./util/config');
 const Cli = require('./util/cli');
 const auth = require('./util/auth');
 const MetadataTypeInfo = require('./MetadataTypeInfo');
@@ -103,7 +104,7 @@ saved
      */
     static async buildTemplate(businessUnit, selectedType, keyArr, market) {
         Util.logger.info('mcdev:: Build Definition from Template');
-        const properties = File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         if (!Util._isValidType(selectedType)) {
             return;
         }
@@ -170,7 +171,7 @@ saved
      * @returns {Promise.<TYPE.MultiMetadataTypeList>} -
      */
     static async buildDefinition(businessUnit, selectedType, name, market) {
-        const properties = File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         if (!Util._isValidType(selectedType)) {
             return;
         }
@@ -197,7 +198,7 @@ saved
      * @returns {Promise.<void>} -
      */
     static async buildDefinitionBulk(listName, type, name) {
-        const properties = File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         Util.verifyMarketList(listName, properties);
         if (type && !MetadataTypeInfo[type]) {
             Util.logger.error(`:: '${type}' is not a valid metadata type`);
