+++ lib/Deployer.js
@@ -6,6 +6,7 @@ const path = require('path');
 const Cli = require('./util/cli');
 const Util = require('./util/util');
 const File = require('./util/file');
+const config = require('./util/config');
 const cache = require('./util/cache');
 const auth = require('./util/auth');
 
@@ -51,7 +52,7 @@ class Deployer {
      */
     static async deploy(businessUnit, selectedTypesArr, keyArr, fromRetrieve) {
         Util.logger.info('mcdev:: Deploy');
-        const properties = File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         if (fromRetrieve) {
             properties.directories.deploy = properties.directories.retrieve;
         }
