+++ lib/index.js
@@ -4,8 +4,8 @@ const TYPE = require('../types/mcdev.d');
 const Util = require('./util/util');
 const auth = require('./util/auth');
 const File = require('./util/file');
+const config = require('./util/config');
 const Init = require('./util/init');
-
 const InitGit = require('./util/init.git');
 const Cli = require('./util/cli');
 const DevOps = require('./util/devops');
@@ -16,7 +16,6 @@ const MetadataTypeInfo = require('./MetadataTypeInfo');
 const MetadataTypeDefinitions = require('./MetadataTypeDefinitions');
 const Retriever = require('./Retriever');
 const cache = require('./util/cache');
-let properties;
 
 /**
  * main class
@@ -55,8 +54,8 @@ class Mcdev {
      */
     static async createDeltaPkg(argv) {
         Mcdev.setSkipInteraction(argv.skipInteraction);
-        properties = properties || File.loadConfigFile();
-        if (!(await Util.checkProperties(properties))) {
+        const properties = await config.loadConfigFile();
+        if (!(await config.checkProperties(properties))) {
             return null;
         }
         // get source market and source BU from config
@@ -72,8 +71,8 @@ class Mcdev {
      * @returns {Promise} .
      */
     static async selectTypes() {
-        properties = properties || File.loadConfigFile();
-        if (!(await Util.checkProperties(properties))) {
+        const properties = await config.loadConfigFile();
+        if (!(await config.checkProperties(properties))) {
             return null;
         }
         await Cli.selectTypes(properties);
@@ -90,7 +89,7 @@ class Mcdev {
      */
     static async upgrade(skipInteraction) {
         Mcdev.setSkipInteraction(skipInteraction);
-        properties = properties || File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         if (!properties) {
             Util.logger.error('No config found. Please run mcdev init');
             return false;
@@ -113,8 +112,8 @@ class Mcdev {
      */
     static async retrieve(businessUnit, selectedTypesArr, keys, changelogOnly) {
         Util.logger.info('mcdev:: Retrieve');
-        properties = properties || File.loadConfigFile();
-        if (!(await Util.checkProperties(properties))) {
+        const properties = await config.loadConfigFile();
+        if (!(await config.checkProperties(properties))) {
             // return null here to avoid seeing 2 error messages for the same issue
             return null;
         }
@@ -202,7 +201,7 @@ class Mcdev {
      * @returns {Promise.<object>} ensure that BUs are worked on sequentially
      */
     static async _retrieveBU(cred, bu, selectedTypesArr, keys, changelogOnly) {
-        properties = properties || File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         const buObject = await Cli.getCredentialObject(
             properties,
             cred !== null ? cred + '/' + bu : null,
@@ -289,7 +288,7 @@ class Mcdev {
     static async initProject(credentialsName, skipInteraction) {
         Util.logger.info('mcdev:: Setting up project');
         Mcdev.setSkipInteraction(skipInteraction);
-        properties = properties || File.loadConfigFile(!!credentialsName);
+        const properties = await config.loadConfigFile(!!credentialsName);
         await Init.initProject(properties, credentialsName, skipInteraction);
     }
 
@@ -301,7 +300,7 @@ class Mcdev {
      */
     static async findBUs(credentialsName) {
         Util.logger.info('mcdev:: Load BUs');
-        properties = properties || File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         const buObject = await Cli.getCredentialObject(properties, credentialsName, true);
         if (buObject !== null) {
             BuHelper.refreshBUProperties(properties, buObject.credential);
@@ -317,7 +316,7 @@ class Mcdev {
      */
     static async document(businessUnit, type) {
         Util.logger.info('mcdev:: Document');
-        properties = properties || File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         if (type && !MetadataTypeInfo[type]) {
             Util.logger.error(`:: '${type}' is not a valid metadata type`);
             return;
@@ -352,7 +351,7 @@ class Mcdev {
      */
     static async deleteByKey(businessUnit, type, customerKey) {
         Util.logger.info('mcdev:: delete');
-        properties = properties || File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         const buObject = await Cli.getCredentialObject(properties, businessUnit);
         if (buObject !== null) {
             if ('string' !== typeof type) {
@@ -381,7 +380,7 @@ class Mcdev {
      * @returns {Promise.<void>} -
      */
     static async badKeys(businessUnit) {
-        properties = properties || File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         const buObject = await Cli.getCredentialObject(properties, businessUnit);
         if (buObject !== null) {
             Util.logger.info('Gathering list of Name<>External Key mismatches (bad keys)');
@@ -452,7 +451,7 @@ class Mcdev {
      */
     static async retrieveAsTemplate(businessUnit, selectedType, name, market) {
         Util.logger.info('mcdev:: Retrieve as Template');
-        properties = properties || File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         if (!Util._isValidType(selectedType)) {
             return;
         }
@@ -526,7 +525,7 @@ class Mcdev {
      */
     static async getFilesToCommit(businessUnit, selectedType, keyArr) {
         Util.logger.info('mcdev:: getFilesToCommit');
-        const properties = File.loadConfigFile();
+        const properties = await config.loadConfigFile();
         if (!Util._isValidType(selectedType)) {
             return;
         }
