+++ lib/util/cli.js
@@ -3,6 +3,7 @@
 const TYPE = require('../../types/mcdev.d');
 const BuHelper = require('./businessUnit');
 const File = require('./file');
+const config = require('./config');
 const inquirer = require('inquirer');
 const MetadataDefinitions = require('./../MetadataTypeDefinitions');
 const Util = require('./util');
@@ -23,7 +24,7 @@ const Cli = {
     async initMcdevConfig(skipInteraction) {
         Util.logger.info('-- Initialising server connection --');
         Util.logger.info('Please enter a name for your "Installed Package" credentials:');
-        const propertiesTemplate = Util.getDefaultProperties();
+        const propertiesTemplate = await config.getDefaultProperties();
         delete propertiesTemplate.credentials.default;
 
         // wait for the interaction to finish or else an outer await will run before this is done
@@ -37,7 +38,7 @@ const Cli = {
      * @returns {Promise.<void>} -
      */
     async addExtraCredential(properties, skipInteraction) {
-        if (!(await Util.checkProperties(properties))) {
+        if (!(await config.checkProperties(properties))) {
             // return null here to avoid seeing 2 error messages for the same issue
             return null;
         } else {
@@ -84,7 +85,7 @@ const Cli = {
      */
     async getCredentialObject(properties, target, isCredentialOnly, allowAll) {
         try {
-            if (!(await Util.checkProperties(properties))) {
+            if (!(await config.checkProperties(properties))) {
                 // return null here to avoid seeing 2 error messages for the same issue
                 return null;
             }
