+++ lib/util/init.config.js
@@ -3,6 +3,7 @@
 const TYPE = require('../../types/mcdev.d');
 const Cli = require('./cli');
 const File = require('./file');
+const config = require('./config');
 const Util = require('./util');
 const inquirer = require('inquirer');
 const path = require('path');
@@ -29,8 +30,8 @@ const Init = {
 
         const upgradeMsgs = [`Upgrading existing ${Util.configFileName}:`];
 
-        const missingFields = await Util.checkProperties(properties, true);
-        const defaultProps = Util.getDefaultProperties();
+        const missingFields = await config.checkProperties(properties, true);
+        const defaultProps = await config.getDefaultProperties();
         if (missingFields.length) {
             missingFields.forEach((fieldName) => {
                 switch (fieldName) {
