+++ lib/util/init.js
@@ -2,6 +2,7 @@
 const TYPE = require('../../types/mcdev.d');
 const Cli = require('./cli');
 const File = require('./file');
+const config = require('./config');
 const InitGit = require('./init.git');
 const InitNpm = require('./init.npm');
 const InitConfig = require('./init.config');
@@ -52,7 +53,7 @@ const Init = {
                     }
                 } while (error && !skipInteraction);
                 Util.logger.debug('reloading config');
-                properties = File.loadConfigFile(true);
+                properties = await config.loadConfigFile(true);
             } else if (missingCredentials.length) {
                 // forced update-credential mode - user likely cloned repo and is missing mcdev-auth.json
                 Util.logger.warn(
@@ -82,7 +83,7 @@ const Init = {
                         }
                     } while (error);
                     Util.logger.debug('reloading config');
-                    properties = File.loadConfigFile(true);
+                    properties = await config.loadConfigFile(true);
                 }
                 Util.logger.info('✔️  All credentials updated.');
                 // assume node dependencies are not installed
